[profile] Fix profile count handling
The profile change notifications were not working in case
the profile count was anything else than the hardcoded number
of 4. Fixed the handling by converting the "names" private as
QStringList and checking against it's size and a editable macro
max value for profile count when filling the profile state arrays.

Increased the maximum profile count from 4 to 10 for increased
robustness.

Modified also unit test cases to work with the QStringList.

Signed-off-by: Kalle Jokiniemi <kalle.jokiniemi@jollamobile.com>