[libprofile-qt] Remove badly working tests. Fixes JB#52781
These seem mostly pointless. The whole project is a thin wrapper on
top of d-bus so basically the test is limited to check that a method
call results in specific kind of d-bus call request.
But then again it can't check the call is made, only stubbing the
internal implementation.

And that's not working, all the tests are crashing. Let's just delete
this.